Rework library so that node ids are not external
This is a breaking change for the library interface. The main change to
API is to stop storing node data/keys and using those are references.
Instead, users will generate new nodes by calling `add_node`, which
will return a id/reference for that node in the topological order. This
node id is used in all the other APIs as well, replacing the previous
generic key parameter.

Other changes include:
 - remove the `bimap` module, as it is no longer needed to maintain
   the node <-> key relationship
 - remove the `failure*` dependencies and writing our own error type
 - Adding the `generational-arena` dependency as the new backing
   store. Using the arena indices prevents the ABA problem.
 - Updating the tests/examples/benchmarks to use the new API.

Downsides to this change are a performance decrease. When benchmarking
the change, I saw up to 300% in regressions. Followup commits will be
working on fixing this regression.