Use fnv hashing instead of default SipHasher
The `fnv` crate advertises "The FNV hash function
 is a custom Hasher implementation that is more
efficient for smaller hash keys." Since most of the HashSet
contents are indexes that are 2 `u64`s in size, this would
be a good match.

Benchmarks show mostly improvements ranging up to 36% faster.
There were one or two regressions in the mutatinig benches,
one that was 22% slower. On the whole, the change is an
improvement.