Replace internal use of arena index with direct index
Previously the internal structures of this crate used the `ArenaIndex`
type for tracking parent/child/visited/etc information. However, this
index type consists of a `(u64, usize)` pair, which is potentially 16
bytes in size on a 64-bit platform.

This PR changes over to use a `UnsafeIndex`, which is just a `usize`
for all internal purposes. The user of the API will still use the
wrapper `(u64, usize)`, but this will be checked and transformed into
the `usize` representation internally.

This optimization is safe because we keep all the references to nodes
fresh in the various tracked locations. There are a couple specific
locations which are important:
  1. The `NodeData.children` or `NodeData.parent` fields. These are
     safe with this optimzation because we remove delete nodes in
     the `delete_node` operation. This ensures that the nodes
     references in this structures will always exist.
  2. The `Descendants` and `DescendantsUnsorted` structures contain
     node references. These are safe because no mutating operation
     can occur while these structures exist, as they hold shared
     (non-mutating) references to the original `IncrementalTopo`
     object.