Fix bug in add_dependency handling of cycle errors
This commit adds a step in the `add_dependency` function which removes
incorrect parent/child information from the successor and predecessor
nodes when a cycle is detected.

Previously this parent/child data was not removed on failure, and was
causing issues in the iteration.

Added an integration test with details and a test which covers the
failure.

Also added more documentation tests which show example of the
`add_dependency` function failing with `Error::CycleDetected`.